# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Koyomi Gal 后端 API",
        "title": "Koyomi Gal API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/galgames/{id}/resources": {
            "get": {
                "description": "返回 Galgame 下全部已发布资源及其链接",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "查看 Galgame 资源列表",
                "operationId": "listGalgameResources",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Galgame ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "资源列表",
                        "schema": {
                            "$ref": "#/definitions/dto.ResourceListResponse"
                        }
                    },
                    "400": {
                        "description": "Galgame ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Galgame 不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "查询资源失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/resources": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "登录用户为已发布 Galgame 上传资源及下载链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "创建资源",
                "operationId": "createResource",
                "parameters": [
                    {
                        "description": "创建资源请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateResourceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/dto.ResourceDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Galgame 不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "创建资源失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/resources/{id}": {
            "get": {
                "description": "按 ID 返回已发布资源及其链接",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "查看资源详情",
                "operationId": "getResource",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "资源 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "资源详情",
                        "schema": {
                            "$ref": "#/definitions/dto.ResourceDataResponse"
                        }
                    },
                    "400": {
                        "description": "资源 ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "资源不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "查询资源失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "全量更新资源字段和链接；上传者本人或拥有 resource:update 权限",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "更新资源",
                "operationId": "updateResource",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "资源 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新资源请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateResourceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/dto.ResourceDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "没有管理该资源的权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "资源不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "更新资源失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除资源及其链接并原子减少计数；上传者本人或拥有 resource:delete 权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "删除资源",
                "operationId": "deleteResource",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "资源 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "资源已删除",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "资源 ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "没有管理该资源的权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "资源不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "删除资源失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateResourceRequest": {
            "type": "object",
            "required": [
                "galgame_id",
                "links",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "官方汉化整合包"
                },
                "galgame_id": {
                    "type": "integer",
                    "example": 1
                },
                "links": {
                    "type": "array",
                    "maxItems": 50,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://example.com/dl",
                        "https://example.com/dl2"
                    ]
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2,
                        3
                    ],
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "千恋＊万花 官方整合包"
                },
                "type": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2,
                        3,
                        4,
                        5,
                        6
                    ],
                    "example": 1
                }
            }
        },
...
        "dto.ResourceData": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "官方汉化整合包"
                },
                "galgame_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ResourceLinkData"
                    }
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "千恋＊万花 官方整合包"
                },
                "type": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string"
                },
                "uploader_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.ResourceDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/dto.ResourceData"
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "dto.ResourceLinkData": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com/dl"
                }
            }
        },
        "dto.ResourceListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ResourceData"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "dto.UpdateResourceRequest": {
            "type": "object",
            "required": [
                "links",
                "status",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "官方汉化整合包"
                },
                "links": {
                    "type": "array",
                    "maxItems": 50,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://example.com/dl"
                    ]
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2,
                        3
                    ],
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "千恋＊万花 官方整合包"
                },
                "type": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2,
                        3,
                        4,
                        5,
                        6
                    ],
                    "example": 1
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 233
                },
                "msg": {
                    "type": "string",
                    "example": "业务错误"
                }
            }
        },
        "response.MessageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "传入 \"Bearer {access_token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}